+++ test/RebindFix.hs
@@ -1,5 +1,5 @@
 {-# LANGUAGE NoImplicitPrelude #-}
-{-# LANGUAGE RebindableSyntax #-}
+{-# LANGUAGE RebindableSyntax  #-}
 -- | RebindableSyntax fixture.
 --
 -- The @if-then-else@ in 'rebindIf' desugars to a call of the local
@@ -7,8 +7,8 @@
 -- @ifThenElse@ in 'rebindIf'.
 module RebindFix (rebindIf) where
 
-import qualified Prelude       as P
-import           Prelude       (Bool (..), Int, fromInteger)
+import           Prelude (Bool (..), Int, fromInteger)
+import qualified Prelude as P
 
 ifThenElse :: Bool -> Int -> Int -> Int
 ifThenElse True  t _ = t P.+ 1000